change variable type from long long to int32_t.

+++ include/backtrack.h
@@ -13,10 +13,10 @@
 
 using namespace std;
 
-extern vector<long long> global_temp_vector;
-extern long long* markNode;
-extern long long* markCell;
-extern long long global_mark;
+extern vector<int32_t> global_temp_vector;
+extern int32_t* markNode;
+extern int32_t* markCell;
+extern int32_t global_mark;
 
 class Backtrack
 {
@@ -24,30 +24,30 @@ class Backtrack
 	Graph* g1 = NULL;
 	Graph* g2 = NULL;
 	Coloring* coloring = NULL;
-	long long numTreeNode = 0;
+	int32_t numTreeNode = 0;
 	//end input
 
 	//output
-	long long n = 0;
-	long long e = 0;
-	long long n2 = 0;
-	long long e2 = 0;
+	int32_t n = 0;
+	int32_t e = 0;
+	int32_t n2 = 0;
+	int32_t e2 = 0;
 	DAG* dag = NULL;
 	CS* cs = NULL;
-	long long* mapping = NULL;
+	int32_t* mapping = NULL;
 
 	//variables used in backtracking
 	// - for the partial failing set
-	vector<long long>* failingset = NULL;
+	vector<int32_t>* failingset = NULL;
 
 	//variables used in backtracking (Workspace)
 	// - for adaptive matching order
-	vector<long long>* extCand = NULL;
+	vector<int32_t>* extCand = NULL;
 	Heap* heap = NULL;
-	long long* weight = NULL;
-	long long* numMappedParent = NULL;
-	long long* candPos = NULL;
-	long long* matchingOrder = NULL;
+	int32_t* weight = NULL;
+	int32_t* numMappedParent = NULL;
+	int32_t* candPos = NULL;
+	int32_t* matchingOrder = NULL;
 	char* isBinary = NULL;
 	
 
@@ -55,30 +55,30 @@ class Backtrack
 	void clearWorkspace();
 
 	DAG* buildDAG();
-	long long selectRoot();
+	int32_t selectRoot();
 
 	CS* buildCS();
-	long long binarySearch(long long*, long long, long long);
+	int32_t binarySearch(int32_t*, int32_t, int32_t);
 
-	long long mapBinaryCell();
+	int32_t mapBinaryCell();
 
-	bool backtrack(long long);
+	bool backtrack(int32_t);
 	//for the adaptive matching order
-	void insertExtVertex(long long, long long);
-	long long getMinExtVertex();
-	void deleteExtVertex(long long);
-	long long computeWeight(long long);
-	void computeExtCand(long long);
+	void insertExtVertex(int32_t, int32_t);
+	int32_t getMinExtVertex();
+	void deleteExtVertex(int32_t);
+	int32_t computeWeight(int32_t);
+	void computeExtCand(int32_t);
 	//for the partial failing set
-	long long binarySearch(vector<long long>&, long long);
-	void merge(vector<long long>&, vector<long long>&);
-	void merge(vector<long long>&, long long*, long long);
+	int32_t binarySearch(vector<int32_t>&, int32_t);
+	void merge(vector<int32_t>&, vector<int32_t>&);
+	void merge(vector<int32_t>&, int32_t*, int32_t);
 	
 public:
 	Backtrack();
 	~Backtrack();
 
-	bool run(Coloring*, Graph*, Graph*, long long);
+	bool run(Coloring*, Graph*, Graph*, int32_t);
 };
 
 #endif
