+++ src/algorithm.cpp
@@ -46,16 +46,16 @@ bool Algorithm::checkSimpleInvariants(Graph* aG1, Graph* aG2)
 	if (aG1->numEdge != aG2->numEdge)
 		return false;
 
-	const long long numNode = aG1->numNode;
+	const int32_t numNode = aG1->numNode;
 
-	long long* nodes1 = global_memory.getLLArray(numNode);
-	long long* nodes2 = global_memory.getLLArray(numNode);
+	int32_t* nodes1 = global_memory.getLLArray(numNode);
+	int32_t* nodes2 = global_memory.getLLArray(numNode);
 
-	for (long long i = 0; i < numNode; ++i)
+	for (int32_t i = 0; i < numNode; ++i)
 		nodes1[i] = nodes2[i] = i;
 
 	sort(nodes1, nodes1 + numNode,
-			[aG1](const long long& i, const long long& j) -> bool {
+			[aG1](const int32_t& i, const int32_t& j) -> bool {
 				if (aG1->d[i] == aG1->d[j])
 					return aG1->l[i] < aG1->l[j];
 
@@ -63,14 +63,14 @@ bool Algorithm::checkSimpleInvariants(Graph* aG1, Graph* aG2)
 			});
 
 	sort(nodes2, nodes2 + numNode,
-			[aG2](const long long& i, const long long& j) -> bool {
+			[aG2](const int32_t& i, const int32_t& j) -> bool {
 				if (aG2->d[i] == aG2->d[j])
 					return aG2->l[i] < aG2->l[j];
 
 				return aG2->d[i] < aG2->d[j];
 			});
 
-	for (long long i = 0; i < numNode; ++i) {
+	for (int32_t i = 0; i < numNode; ++i) {
 		if (aG1->d[nodes1[i]] != aG2->d[nodes2[i]])
 			return false;
 		if (aG1->l[nodes1[i]] != aG2->l[nodes2[i]])
